src/sage/calculus/all.py: Move 'symbolic_expression' to new module .expr
